changed location of global sockaddr_in
